Order details blade: fixed order details disappearing from the order operations tree after editing the payment details for the order.
