+++ VirtoCommerce.OrderModule.Data/Model/OperationEntity.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
@@ -8,6 +8,7 @@ using System.Reflection;
 using Omu.ValueInjecter;
 using VirtoCommerce.Domain.Commerce.Model;
 using VirtoCommerce.Domain.Order.Model;
+using VirtoCommerce.OrderModule.Data.Utilities;
 using VirtoCommerce.Platform.Core.Common;
 
 namespace VirtoCommerce.OrderModule.Data.Model
@@ -40,7 +41,7 @@ namespace VirtoCommerce.OrderModule.Data.Model
 
             operation.InjectFrom(this);
 
-            operation.ChildrenOperations = GetAllChildOperations(operation);
+            operation.ChildrenOperations = OperationUtilities.GetAllChildOperations(operation);
             return operation;
         }
 
@@ -70,38 +71,6 @@ namespace VirtoCommerce.OrderModule.Data.Model
             operation.CancelReason = CancelReason;
             operation.IsApproved = IsApproved;
             operation.Sum = Sum;
-        }
-
-        private static IEnumerable<IOperation> GetAllChildOperations(IOperation operation)
-        {
-            var retVal = new List<IOperation>();
-            var objectType = operation.GetType();
-
-            var properties = objectType.GetProperties(BindingFlags.Public | BindingFlags.Instance);
-
-            var childOperations = properties.Where(x => x.PropertyType.GetInterface(typeof(IOperation).Name) != null)
-                                    .Select(x => (IOperation)x.GetValue(operation)).Where(x => x != null).ToList();
-
-            foreach (var childOperation in childOperations)
-            {
-                retVal.Add(childOperation);
-            }
-
-            //Handle collection and arrays
-            var collections = properties.Where(p => p.GetIndexParameters().Length == 0)
-                                        .Select(x => x.GetValue(operation, null))
-                                        .Where(x => x is IEnumerable && !(x is string))
-                                        .Cast<IEnumerable>();
-
-            foreach (var collection in collections)
-            {
-                foreach (var childOperation in collection.OfType<IOperation>())
-                {
-                    retVal.Add(childOperation);
-                }
-            }
-
-            return retVal;
         }
     }
 }
